fix(giskard-checks): quality fixes for BaseLLMGenerator and LLMGenerator
- Remove @override from BaseLLMGenerator.__call__ (async generator protocol differs)
- Align LLMGenerator validator error messages to match LLMJudge style (hyphen, no trailing period)
- Remove dead second response in test_llm_generator_stops_at_max_steps
- Add tests for max_steps=0 and message=None-with-goal-not-reached cases
- Extract MockGenerator and LLMTrace to shared conftest.py; add __init__.py to enable relative imports
